Add FilePickerAction for file selection functionality
- Introduced FilePickerAction to handle file selection from gallery or files.
- Updated ScreenController to remove deprecated file manager references.
- Registered FilePickerAction in ActionInvokable for action handling.
- Adjusted imports in relevant files to integrate the new action.
