+++ modules/ensemble/lib/action/action_invokable.dart
@@ -45,6 +45,7 @@ abstract class ActionInvokable with Invokable {
       ActionType.controlDeviceBackNavigation,
       ActionType.closeApp,
       ActionType.getLocation,
+      ActionType.pickFiles,
     ]);
   }
