+++ modules/ensemble/lib/framework/stub/file_manager.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import 'package:ensemble/framework/action.dart';
+import 'package:ensemble/action/file_picker_action.dart';
 import 'package:ensemble/framework/error_handling.dart';
 import 'package:ensemble/framework/scope.dart';
 import 'package:flutter/cupertino.dart';
